# ccipreader optimization - sort and limit